moved hydroshare oauth route handlers into dataloaderinterface; updated hydroshare model; fixed some small bugs
