+++ src/dataloaderinterface/urls.py
@@ -17,7 +17,7 @@ from django.conf.urls import url
 
 from dataloaderinterface.views import SitesListView, SiteDetailView, SiteRegistrationView, \
     HomeView, BrowseSitesListView, SiteUpdateView, SiteDeleteView, StatusListView, HydroShareResourceUpdateView, \
-    HydroShareResourceCreateView
+    HydroShareResourceCreateView, OAuthAuthorize, OAuthAuthorizeRedirect
 
 urlpatterns = [
     url(r'^$', HomeView.as_view(), name='home'),
@@ -30,4 +30,6 @@ urlpatterns = [
     url(r'^sites/(?P<sampling_feature_code>.*)/hsr/create/$', HydroShareResourceCreateView.as_view(), name='hs_resource_create'),
     url(r'^sites/(?P<sampling_feature_code>.*)/hsr/update/$', HydroShareResourceUpdateView.as_view(), name='hs_resource_update'),
     url(r'^sites/(?P<sampling_feature_code>.*)/$', SiteDetailView.as_view(), name='site_detail'),
+    url(r'hydroshare/oauth/', OAuthAuthorize.as_view(), name='hydroshare_oauth'),
+    url(r'hydroshare/oauth/redirect/', OAuthAuthorizeRedirect.as_view(), name='hydroshare_oauth_redirect'),
 ]
